[1898774] Added  API to remove client user by admin

See merge request clientprojects/assembler_projects/prd/londonfilmedeventsfinalbuild-390988-ruby!48

+++ template-app/app/controllers/account_block/accounts_controller.rb
@@ -7,8 +7,8 @@ module AccountBlock
     # before_action :validate_json_web_token, only: [:search, :change_email_address, :change_phone_number, :specific_account, :logged_user, :change_password, :update, :add_client_user]
 
     # before_action :current_user, only: [:change_password, :update, :add_client_user]
-    before_action :validate_client_admin, only: [:add_client_user, :update_client_user, :client_users, :remove_user, :company_users]
-    before_action :validate_client_admin_permission, only: [:add_client_user, :update_client_user, :remove_user]
+    before_action :validate_client_admin, only: [:add_client_user, :update_client_user, :client_users, :remove_user, :company_users, :send_account_activation_email]
+    before_action :validate_client_admin_permission, only: [:add_client_user, :update_client_user, :remove_user, :send_account_activation_email]
 
     def create
       case params[:data][:type] #### rescue invalid API format
@@ -240,12 +240,20 @@ module AccountBlock
     end
 
     def remove_user
-      client_user = @account.client_users.find_by(id: params[:user_id])
+      client_user = @account.company.accounts.find_by_id(params[:user_id]) rescue nil
+      return render json: {message: "User not present or you're not authorized to delete this user"}, status: :unauthorized unless client_user.present?
       if client_user && client_user.destroy
         render json: { message: "Client user removed successfully" }, status: :ok
       else
         render json: { message: "Unable to remove client user" }, status: :unprocessable_entity
       end
+    end
+
+    def send_account_activation_email
+      client_user = @account.company.accounts.find_by_id(params[:user_id]) rescue nil
+      return render json: {message: "User not present or you're not authorized"}, status: :unauthorized unless client_user.present?
+      EmailValidationMailer.with(account: client_user, host: request.base_url).activation_email.deliver
+      render json: {message: "Success"}, status: :ok
     end
 
     def logged_user
